src/deepfake-video/deepfake_video/video_detector/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
argcomplete==3.5.1
astor==0.8.1
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cmake==3.30.5
datamodel-code-generator==0.26.2
dlib==19.24.6
dnspython==2.7.0
email_validator==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
flask_ml==0.2.5
fsspec==2024.10.0
gast==0.6.0
genson==1.3.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
inflect==5.6.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.6.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
nvidia-ml-py3==7.352.0
olefile==0.47
opencv-python==4.10.0.84
optree==0.13.0
packaging==24.1
pathlib==1.0.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
protobuf==5.28.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scandir==1.10.0
six==1.16.0
sympy==1.13.1
termcolor==2.5.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
